requirements.txt

langchain-openai
langchain-groq
langchain_community
langchain-text-splitters
pypdf
pandas

langchain==1.2.0
python-dotenv==1.2.1
ipykernel==7.1.0
langchain_groq==1.1.1
langchain-google-genai==4.1.2
langchain-community==0.4.1
faiss-cpu==1.13.2
structlog==25.5.0
PyMuPDF==1.26.7
langchain-core==1.2.6
streamlit==1.53.0
docx2txt==0.9
fastapi
uvicorn
langchain-huggingface
python-multipart
sentence-transformers


langchain-google-genai

-e .